GENAI_Model/requirements.txt
#python-dotenv
google-generativeai
streamlit
configparser
torch
psycopg
ultralytics
boto3
opencv-python-headless


